+++ fe/fe-core/src/main/java/com/starrocks/planner/stream/StreamAggNode.java
@@ -21,7 +21,7 @@ import com.starrocks.common.FeConstants;
 import com.starrocks.planner.FragmentNormalizer;
 import com.starrocks.planner.PlanNode;
 import com.starrocks.planner.PlanNodeId;
-import com.starrocks.sql.optimizer.operator.physical.stream.IMTInfo;
+import com.starrocks.sql.optimizer.operator.stream.IMTInfo;
 import com.starrocks.thrift.TExplainLevel;
 import com.starrocks.thrift.TExpr;
 import com.starrocks.thrift.TNormalPlanNode;
