+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/stream/IMTInfo.java
@@ -1,0 +1,7 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
+
+package com.starrocks.sql.optimizer.operator.stream;
+
+// TODO
+public class IMTInfo {
+}
