+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/stream/PhysicalStreamJoinOperator.java
@@ -1,19 +1,6 @@
-// Copyright 2021-present StarRocks, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 
-
-package com.starrocks.sql.optimizer.operator.physical.stream;
+package com.starrocks.sql.optimizer.operator.stream;
 
 import com.starrocks.analysis.JoinOperator;
 import com.starrocks.sql.optimizer.OptExpression;
@@ -53,7 +40,6 @@ public class PhysicalStreamJoinOperator extends PhysicalStreamOperator {
     public String getJoinHint() {
         return joinHint;
     }
-
 
     @Override
     public <R, C> R accept(OperatorVisitor<R, C> visitor, C context) {
